mgmt_system fixtures have a new home
- They're in the fixtures dir now
